requirements.txt
click==5.1
Django==1.9
django-appconf==1.0.1
django-bower==5.0.4
django-compressor==1.6
django-extensions==1.5.9
djangorestframework==3.3.2
gunicorn==19.2.1
Jinja2==2.8
livereload==2.4.0
Markdown==2.6.2
MarkupSafe==0.23
mkdocs==0.14.0
psycopg2==2.6.1
PyYAML==3.11
six==1.9.0
tornado==4.2.1
Werkzeug==0.10.4
wheel==0.24.0
South==1.0.2